requirements.txt
fastapi
uvicorn
beautifulsoup4
requests
pymysql
sqlalchemy
aioredis
